requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
arabic-reshaper==3.0.0
asn1crypto==1.5.1
attrs==25.3.0
azure-ai-projects==1.0.0b10
azure-core==1.34.0
azure-functions==1.23.0
azure-identity==1.22.0
azure-storage-blob==12.25.1
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.0
colorama==0.4.6
cryptography==44.0.3
cssselect2==0.8.0
defusedxml==0.7.1
distro==1.9.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
idna==3.10
isodate==0.7.2
Jinja2==3.1.6
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
langchain-core==0.3.59
langchain-openai==0.3.16
langsmith==0.3.42
lxml==5.4.0
MarkupSafe==3.0.2
msal==1.32.3
msal-extensions==1.3.1
narwhals==1.39.0
numpy==2.2.5
openai==1.78.1
orjson==3.10.18
oscrypto==1.3.0
packaging==24.2
pandas==2.2.3
pillow==11.2.1
protobuf==6.30.2
pyarrow==20.0.0
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
pydeck==0.9.1
pyHanko==0.26.0
pyhanko-certvalidator==0.26.8
PyJWT==2.10.1
PyMuPDF==1.25.5
pyodbc==5.2.0
pypdf==5.5.0
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
qrcode==8.2
referencing==0.36.2
regex==2024.11.6
reportlab==4.4.0
requests==2.32.3
requests-toolbelt==1.0.0
rpds-py==0.24.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
streamlit==1.45.0
svglib==1.5.1
tenacity==9.1.2
tiktoken==0.9.0
tinycss2==1.4.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.4.0
watchdog==6.0.0
webencodings==0.5.1
Werkzeug==3.1.3
xhtml2pdf==0.2.17
zstandard==0.23.0
